♻️ refactor:  リスト表示とマップ表示の処理共通化
#74
