+++ lib/presentation/page/home/component/entry_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../../stamp_rally/component/stamp_rally_display_mode.dart';
-import 'entry_details_view.dart';
+import 'entry_list_view.dart';
 import 'entry_map_view.dart';
 
 /// 参加中画面
@@ -15,7 +15,7 @@ class EntryView extends StatelessWidget {
       builder: (mode) {
         switch (mode) {
           case StampRallyDisplayMode.list:
-            return const EntryDetailsView();
+            return const EntryListView();
           case StampRallyDisplayMode.map:
             return const EntryMapView();
         }
